requirements.txt
streamlit
openai
pandas
python-dotenv
streamlit_authenticator
pyyaml
bcrypt